Replacing the matplotlib library dependency (#77)
* lsmop/mafs

* update

* maf11

* complete

* maf check

* finish maf5 check

* finish maf9 check

* finish maf10 check

* finish maf11 check

* fori_loop

* fori_loop

* add ax.disable_jit()

* maf10 JIT

* maf13 JIT finish

* JIT finish

* maf LSMOP finish

* Modify format

* Modified to meet merger requirements. But still need change lsmop.py because of for loop, remove chex for type hint and redundant function liking Sphere.

* Modified to meet merger requirements of "for" in lsmop.py. Still need to remove chex for type hint and redundant function liking Sphere.

* remove chex for type hint

* Remove redundant function in lsmop.py, remove "global" in two test files, fix the bug in griewank.py

* Modify “import” format

* Modify key format

* Remove unnecessary comment

* Change double for-loop nesting to one layer for-loop embedding fori_ Loop

* Reduce redundant code, use double for-loops and fix other problems

* Reduce redundant code

* fix: follow numpy's style

* dev: simplify code

* dev: refactor code

* fix a liitle problem.

* fix a liitle problem, add more test problems for lsmop

* Remove unnecessary comments

* remove matplotlib dependency in MaF8, MaF9 and the "jax.jit_disable" in MaF2

* change the method of test a point in a convex polygon from Angle Sum Method to Ray Casting Method

* change the test problem to avoid the problem of dividing by zero

* remove the global config of float64

* fix bug of gym

* fix bug of gym

* dev: optimize point_in_polygon

* fix: point in polygon and disable explicit float64

* fix: point_in_polygon and add related tests

* fix: typo

* fix: implementation of inside

fix: typo

* fix: typo

* fix: inequation in ray_intersect_segment

* fix bug of inside function in maf.py

* fix: test cases

* change the style of code

* fix: edge cases

* fix: point_in_polygon

fix: computation order

* fix: point is one of the vertex

---------

+++ src/evox/problems/numerical/lsmop.py
@@ -451,7 +451,7 @@ class LSMOP9(LSMOP):
         return p, state
 
     def _grid(self, N, M):
-        gap = jnp.linspace(0, 1, int(math.ceil(N ** (1 / M))), dtype=jnp.float64)
+        gap = jnp.linspace(0, 1, int(math.ceil(N ** (1 / M))))
         c = jnp.meshgrid(*([gap] * M))
         w = jnp.vstack([x.ravel() for x in c]).T
         return w
